Reorg dumb_analyzer module
