+++ crates/maple_cli/src/dumb_analyzer/find_usages/search_engine/ctags.rs
@@ -1,0 +1,1 @@
+fn hello() {}
